kafka added to implement outbox pattern

+++ BeerDispancer/BeerDispencer.WebUi/ViewModels/Response/Dispenser.cs
@@ -4,7 +4,7 @@ namespace BeerDispenser.WebUi.ViewModels.Response
 	public class Dispenser
     {
         public Guid Id { get; set; }
-        public double FlowVolume { get; set; }
+        public decimal FlowVolume { get; set; }
 		
     };
 }
