+++ BeerDispancer/BeerDispencer.WebUi/ViewModels/Response/DispenserWithStatus.cs
@@ -7,7 +7,7 @@ namespace BeerDispenser.WebUi.ViewModels.Response
     public class DispenserWithStatus
 	{
         public Guid Id { get; set; }
-        public double Volume { get; set; }
+        public decimal Volume { get; set; }
         [JsonProperty("status")]
         public DispenserStatus Status { get; set; }
         public string ReservedFor { get; set; }
